@@ -13,25 +13,26 @@
     "tts",
     "text-to-speech"
   ],
-  "homepage": "https://cboard-org.github.io/cboard-translate",
+  "homepage": "https://app.cboard.io",
   "private": false,
   "license": "GPL-3.0-only",
   "dependencies": {
     "@cospired/i18n-iso-languages": "^2.2.0",
-    "@crowdin/crowdin-api-client": "^1.11.1",
+    "@crowdin/crowdin-api-client": "^1.18.2",
+    "@ctrl/react-adsense": "^1.5.0",
     "@material-ui/core": "^4.12.1",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "4.0.0-alpha.57",
-    "@microsoft/applicationinsights-web": "^2.6.3",
-    "@redux-beacon/google-analytics-gtag": "^1.0.2",
+    "@microsoft/applicationinsights-web": "^2.8.4",
+    "@redux-beacon/google-analytics-gtag": "^1.1.0",
     "@redux-beacon/logger": "^1.0.0",
     "@redux-beacon/offline-web": "^1.0.0",
-    "axios": "^0.21.3",
+    "axios": "^0.26.1",
     "browser-image-resizer": "^1.2.0",
     "dom-to-image": "^2.6.0",
     "dotenv": "^8.6.0",
-    "echarts": "4.9.0",
-    "echarts-for-react": "^2.0.15-beta.1",
+    "echarts": "5.3.2",
+    "echarts-for-react": "^3.0.2",
     "file-saver": "^2.0.5",
     "fontsource-roboto": "^4.0.0",
     "formik": "^1.3.2",
@@ -40,58 +41,59 @@
     "intl": "^1.2.5",
     "is-url": "1.2.4",
     "ismobilejs": "^0.5.1",
-    "iso-639-1": "2.1.9",
-    "jss": "^10.7.1",
+    "iso-639-1": "2.1.15",
+    "jss": "^10.9.0",
     "jss-rtl": "^0.3.0",
-    "jszip": "^3.7.0",
+    "jszip": "^3.9.1",
     "jszip-utils": "^0.1.0",
     "keycode": "^2.2.0",
     "lodash": "^4.17.21",
     "mathjs": "7.6.0",
-    "microsoft-cognitiveservices-speech-sdk": "^1.17.0",
-    "mime-types": "^2.1.31",
-    "moment": "2.27.0",
-    "mongoose": "^5.13.0",
+    "microsoft-cognitiveservices-speech-sdk": "^1.21.1",
+    "mime-types": "^2.1.35",
+    "moment": "2.29.4",
+    "mongoose": "^6.4.0",
     "opencollective": "^1.0.3",
-    "pdfmake": "^0.2.0",
-    "prop-types": "^15.7.2",
+    "pdfmake": "^0.2.5",
+    "prop-types": "^15.8.1",
     "query-string": "^6.14.1",
-    "ramda": "^0.27.1",
+    "ramda": "^0.27.2",
     "react": "^17.0.2",
     "react-autosuggest": "^9.4.2",
+    "react-cropper": "^2.1.8",
     "react-dnd": "^11.1.3",
     "react-dnd-touch-backend": "^11.1.3",
     "react-dom": "^17.0.2",
     "react-grid-layout": "^0.16.6",
     "react-helmet": "^6.1.0",
+    "react-icons": "^4.2.0",
     "react-intl": "^2.7.2",
-    "react-joyride": "^2.3.0",
+    "react-joyride": "^2.4.0",
     "react-markdown": "^4.3.1",
     "react-media-recorder": "^0.7.1",
     "react-redux": "^5.1.1",
-    "react-router-dom": "^5.1.2",
+    "react-router-dom": "^5.3.3",
     "react-scannable": "0.0.17",
     "react-share": "^2.3.1",
     "react-sizeme": "^2.5.2",
     "react-social-login-buttons": "^3.5.0",
     "react-transition-group": "4.4.2",
-    "redux": "^4.1.0",
+    "redux": "^4.2.0",
     "redux-beacon": "^2.0.3",
     "redux-logger": "4.0.0",
     "redux-persist": "^5.10.0",
     "redux-thunk": "^2.3.0",
     "shortid": "^2.2.16",
     "source-map-explorer": "^2.1.2",
-    "swiper": "^6.7.5",
+    "swiper": "^6.8.4",
     "uuid": "^3.4.0",
-    "yup": "^0.29.2"
+    "yup": "^0.32.11"
   },
   "devDependencies": {
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.0",
     "decompress-zip": "^0.3.1",
     "enzyme": "^3.11.0",
     "enzyme-to-json": "3.3.5",
-    "gh-pages": "^3.2.3",
     "husky": "^1.1.4",
     "jest-mock-axios": "^3.2.0",
     "lint-staged": "^10.0.3",
@@ -102,7 +104,7 @@
     "react-test-renderer": "^17.0.2",
     "redux-mock-store": "^1.5.4",
     "sw-precache": "^5.2.1",
-    "typescript": "^3.4.2"
+    "typescript": "^4.6.4"
   },
   "husky": {
     "hooks": {
