@@ -33,17 +33,15 @@ export class LanguageProvider extends Component {
     const { lang } = this.props;
 
     if (lang) {
-      this.fetchMessages('zu-ZA');
+      this.fetchMessages(lang);
     } else {
       this.fetchMessages(DEFAULT_LANG);
     }
   }
 
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    const { lang } = nextProps;
-
-    if (lang) {
-      this.fetchMessages('zu-ZA');
+  componentDidUpdate(prevProps) {
+    if (prevProps.lang !== this.props.lang) {
+      this.fetchMessages(this.props.lang);
     }
   }
 
@@ -54,6 +52,8 @@ export class LanguageProvider extends Component {
       setLangs,
       showNotification
     } = this.props;
+
+    //reset state. This is necessary due that messages is a file import
     this.setState({ messages: null });
 
     importTranslation(lang)
@@ -72,6 +72,7 @@ export class LanguageProvider extends Component {
 
   render() {
     const { lang, children } = this.props;
+
     const locale = lang.slice(0, 2);
 
     if (!this.state.messages) {
