@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Switch, Redirect } from 'react-router-dom';
 import classNames from 'classnames';
 
 import Activate from '../Account/Activate';
@@ -31,17 +31,28 @@ export class App extends Component {
      */
     isLogged: PropTypes.bool,
     /**
+     * If 'true', user is downloading a new lang
+     */
+    isDownloadingLang: PropTypes.bool,
+    /**
      * App language
      */
     lang: PropTypes.string.isRequired,
     /**
      * If 'true', dark mode is enabled
      */
-    dark: PropTypes.bool,
+    dark: PropTypes.bool
   };
 
   render() {
-    const { lang, dir, isFirstVisit, isLogged, dark } = this.props;
+    const {
+      lang,
+      dir,
+      isFirstVisit,
+      isLogged,
+      dark,
+      isDownloadingLang
+    } = this.props;
 
     return (
       <div className={classNames('App', { 'is-dark': dark })}>
@@ -50,7 +61,6 @@ export class App extends Component {
         </Helmet>
 
         <Notifications />
-
         <Switch>
           <RedirectIfLogged
             component={AuthScreen}
@@ -64,6 +74,11 @@ export class App extends Component {
           <Route path="/reset/:userid/:url" component={ChangePassword} />
           <Route path="/login/:type/callback" component={OAuthLogin} />
           <Route path="/board/:id" component={BoardContainer} />
+          {isDownloadingLang && (
+            <Route exact path={'/'}>
+              <Redirect to={'/settings/language'} />
+            </Route>
+          )}
           <Route
             exact
             path="/"
