@@ -44,7 +44,9 @@ import {
   updateApiObjects,
   updateApiObjectsNoChild,
   getApiObjects,
-  downloadImages
+  downloadImages,
+  createApiBoard,
+  upsertApiBoard
 } from './Board.actions';
 import {
   upsertCommunicator,
@@ -61,9 +63,9 @@ import {
   SCANNING_METHOD_MANUAL
 } from '../Settings/Scanning/Scanning.constants';
 import { NOTIFICATION_DELAY } from '../Notifications/Notifications.constants';
-import { isAndroid } from '../../cordova-util';
 import { EMPTY_VOICES } from '../../providers/SpeechProvider/SpeechProvider.constants';
 import { DEFAULT_ROWS_NUMBER, DEFAULT_COLUMNS_NUMBER } from './Board.constants';
+//import { isAndroid } from '../../cordova-util';
 
 const Transition = React.forwardRef(function Transition(props, ref) {
   return <Slide direction="up" ref={ref} {...props} />;
@@ -170,7 +172,8 @@ export class BoardContainer extends Component {
     downloadImages: PropTypes.func,
     lang: PropTypes.string,
     isRootBoardTourEnabled: PropTypes.bool,
-    disableTour: PropTypes.func
+    disableTour: PropTypes.func,
+    isLiveMode: PropTypes.bool
   };
 
   state = {
@@ -184,8 +187,15 @@ export class BoardContainer extends Component {
     isGettingApiObjects: false,
     copyPublicBoard: false,
     blockedPrivateBoard: false,
-    isFixedBoard: false
+    isFixedBoard: false,
+    copiedTiles: [],
+    isScroll: false,
+    totalRows: null
   };
+  constructor(props) {
+    super(props);
+    this.boardRef = React.createRef();
+  }
 
   async componentDidMount() {
     const {
@@ -196,13 +206,12 @@ export class BoardContainer extends Component {
 
     const {
       board,
-      boards,
       communicator,
       changeBoard,
       userData,
       history,
-      getApiObjects,
-      downloadImages
+      getApiObjects
+      //downloadImages
     } = this.props;
 
     // Loggedin user?
@@ -215,6 +224,7 @@ export class BoardContainer extends Component {
       this.setState({ isGettingApiObjects: false });
     }
 
+    const boards = this.props.boards; //see board from redux state after get ApiObjets
     let boardExists = null;
 
     if (id && board && id === board.id) {
@@ -271,7 +281,7 @@ export class BoardContainer extends Component {
     //set board type
     this.setState({ isFixedBoard: !!boardExists.isFixed });
 
-    if (isAndroid()) downloadImages();
+    // if (isAndroid()) downloadImages();
   }
 
   UNSAFE_componentWillReceiveProps(nextProps) {
@@ -293,6 +303,7 @@ export class BoardContainer extends Component {
         ) {
           changeBoard(nextProps.match.params.id);
           previousBoard();
+          this.scrollToTop();
         }
       } else {
         // Was a browser back action?
@@ -588,7 +599,7 @@ export class BoardContainer extends Component {
     this.toggleSelectMode();
   };
 
-  handleAddTileEditorSubmit = tile => {
+  handleAddTileEditorSubmit = async tile => {
     const {
       userData,
       createTile,
@@ -612,6 +623,7 @@ export class BoardContainer extends Component {
       createBoard(boardData);
       addBoardCommunicator(boardData.id);
     }
+
     if (tile.type !== 'board') {
       this.updateIfFeaturedBoard(board);
       createTile(tile, board.id);
@@ -619,9 +631,10 @@ export class BoardContainer extends Component {
 
     // Loggedin user?
     if ('name' in userData && 'email' in userData) {
-      this.handleApiUpdates(tile);
+      await this.handleApiUpdates(tile);
       return;
     }
+
     //if not and is adding an emptyBoard
     if (tile.type === 'board') {
       switchBoard(boardData.id);
@@ -735,7 +748,7 @@ export class BoardContainer extends Component {
   };
 
   handleLayoutChange = (currentLayout, layouts) => {
-    const { updateBoard, replaceBoard, board } = this.props;
+    const { updateBoard, replaceBoard, board, navigationSettings } = this.props;
     currentLayout.sort((a, b) => {
       if (a.y === b.y) {
         return a.x - b.x;
@@ -747,10 +760,26 @@ export class BoardContainer extends Component {
 
     const tilesIds = currentLayout.map(gridTile => gridTile.i);
     const tiles = tilesIds.map(t => {
-      return board.tiles.find(
-        tile => tile.id === t || Number(tile.id) === Number(t)
-      );
+      return board.tiles.find(tile => {
+        if (!tile) {
+          return false;
+        }
+        return tile.id === t || Number(tile.id) === Number(t);
+      });
     });
+
+    if (navigationSettings.bigScrollButtonsActive) {
+      const cols =
+        currentLayout.reduce(function(valorAnterior, item) {
+          if (item.x > valorAnterior) return item.x;
+          return valorAnterior;
+        }, 0) + 1;
+      const rows = 3;
+      const isScroll = currentLayout.length / cols > rows ? true : false;
+      const totalRows = Math.ceil(currentLayout.length / cols);
+      this.setIsScroll(isScroll, totalRows);
+    }
+
     const newBoard = { ...board, tiles };
     replaceBoard(board, newBoard);
     if (!isEqual(board.tiles, tiles)) {
@@ -758,6 +787,10 @@ export class BoardContainer extends Component {
       updateBoard(processedBoard);
       this.saveApiBoardOperation(processedBoard);
     }
+  };
+
+  setIsScroll = (bool, totalRows = 0) => {
+    this.setState({ isScroll: bool, totalRows: totalRows });
   };
 
   handleTileDrop = async (tile, position) => {
@@ -815,7 +848,8 @@ export class BoardContainer extends Component {
       intl,
       boards,
       showNotification,
-      navigationSettings
+      navigationSettings,
+      isLiveMode
     } = this.props;
     const hasAction = tile.action && tile.action.startsWith('+');
 
@@ -846,9 +880,21 @@ export class BoardContainer extends Component {
         showNotification(intl.formatMessage(messages.boardMissed));
       }
     } else {
-      changeOutput([...this.props.output, tile]);
       clickSymbol(tile.label);
       say();
+      if (isLiveMode) {
+        const liveTile = {
+          backgroundColor: 'rgb(255, 241, 118)',
+          id: shortid.generate(),
+          image: '',
+          label: '',
+          labelKey: '',
+          type: 'live'
+        };
+        changeOutput([...this.props.output, tile, liveTile]);
+      } else {
+        changeOutput([...this.props.output, tile]);
+      }
     }
   };
 
@@ -967,7 +1013,6 @@ export class BoardContainer extends Component {
       switchBoard,
       lang
     } = this.props;
-
     // Loggedin user?
     if ('name' in userData && 'email' in userData) {
       this.setState({
@@ -1002,7 +1047,9 @@ export class BoardContainer extends Component {
         );
       }
       if (tile && tile.type !== 'board') {
-        uTiles = [...board.tiles, tile];
+        uTiles = board.tiles.find(t => t.id === tile.id)
+          ? [...board.tiles]
+          : [...board.tiles, tile];
       }
       if (tile && tile.type === 'board') {
         uTiles = [...board.tiles];
@@ -1128,6 +1175,21 @@ export class BoardContainer extends Component {
         this.props.navHistory.length - 2
       ];
       this.props.history.replace(`/board/${prevBoardId}`);
+      this.scrollToTop();
+    }
+  }
+
+  onRequestToRootBoard() {
+    this.props.toRootBoard();
+    this.scrollToTop();
+  }
+
+  scrollToTop() {
+    if (this.boardRef && !this.state.isSelecting) {
+      const boardComponentRef = this.props.board.isFixed
+        ? 'fixedBoardContainerRef'
+        : 'boardContainerRef';
+      this.boardRef.current[boardComponentRef].current.scrollTop = 0;
     }
   }
 
@@ -1141,6 +1203,7 @@ export class BoardContainer extends Component {
       showNotification(intl.formatMessage(messages.boardCopyError));
     }
   };
+
   async createBoardsRecursively(board, records) {
     const {
       createBoard,
@@ -1327,6 +1390,138 @@ export class BoardContainer extends Component {
     this.saveApiBoardOperation(processedBoard);
   };
 
+  handleCopyTiles = () => {
+    const { intl, showNotification } = this.props;
+    const copiedTiles = this.selectedTiles();
+    this.setState({
+      copiedTiles: copiedTiles
+    });
+    showNotification(intl.formatMessage(messages.tilesCopiedSuccessfully));
+  };
+
+  handlePasteTiles = async () => {
+    const { board, intl, createTile, showNotification } = this.props;
+    try {
+      this.setState({ isSaving: true });
+      for await (const tile of this.state.copiedTiles) {
+        const newTile = {
+          ...tile,
+          id: shortid.generate()
+        };
+        if (tile.loadBoard) {
+          createTile(newTile, board.id);
+          await this.pasteBoardsRecursively(newTile, board.id);
+        } else {
+          await this.handleAddTileEditorSubmit(newTile);
+        }
+      }
+      showNotification(intl.formatMessage(messages.tilesPastedSuccessfully));
+    } catch (err) {
+      showNotification(intl.formatMessage(messages.tilesPastedError));
+      console.error(err.message);
+    } finally {
+      this.setState({ isSaving: false });
+    }
+  };
+
+  async pasteBoardsRecursively(folderTile, parentBoardId) {
+    const {
+      createBoard,
+      userData,
+      updateBoard,
+      createApiBoard,
+      boards,
+      intl
+    } = this.props;
+
+    //prevent shit
+    if (!folderTile || !folderTile.loadBoard) {
+      return;
+    }
+
+    let newBoard = {
+      ...boards.find(b => b.id === folderTile.loadBoard),
+      isPublic: false,
+      id: shortid.generate(),
+      hidden: false,
+      author: '',
+      email: ''
+    };
+    if (!newBoard.name) {
+      newBoard.name = newBoard.nameKey
+        ? intl.formatMessage({ id: newBoard.nameKey })
+        : intl.formatMessage(messages.noTitle);
+    }
+    if ('name' in userData && 'email' in userData) {
+      newBoard = {
+        ...newBoard,
+        author: userData.name,
+        email: userData.email
+      };
+    }
+    // Prevent creating a board without the tiles property
+    if (newBoard.tiles) createBoard(newBoard);
+    // Loggedin user?
+    if ('name' in userData && 'email' in userData) {
+      try {
+        newBoard = await createApiBoard(newBoard, newBoard.id);
+      } catch (err) {
+        console.error(err.message);
+      }
+    }
+    const parentBoard = boards.find(b => b.id === parentBoardId);
+    const newTiles = parentBoard.tiles.map(tile =>
+      tile && tile.id === folderTile.id
+        ? { ...tile, loadBoard: newBoard.id }
+        : tile
+    );
+    const boardData = { ...parentBoard, tiles: newTiles };
+    updateBoard(boardData);
+    // Loggedin user?
+    if ('name' in userData && 'email' in userData) {
+      try {
+        let newParentBoard = {
+          ...boardData,
+          hidden: false,
+          author: userData.name,
+          email: userData.email
+        };
+        if (!newParentBoard.name) {
+          newParentBoard.name = newParentBoard.nameKey
+            ? intl.formatMessage({ id: newParentBoard.nameKey })
+            : intl.formatMessage(messages.noTitle);
+        }
+        await this.handleApiUpdates('', '', '', newParentBoard);
+      } catch (err) {
+        console.error(err.message);
+      }
+    }
+
+    //return condition
+    newBoard.tiles.forEach(async tile => {
+      if (tile && tile.loadBoard) {
+        //look for this board in available boards
+        const newBoardToCopy = boards.find(b => b.id === tile.loadBoard);
+        if (newBoardToCopy) {
+          this.pasteBoardsRecursively(tile, newBoard.id);
+        }
+      }
+    });
+    return;
+  }
+
+  selectedTiles = () => {
+    return this.state.selectedTileIds
+      ? this.state.selectedTileIds.map(selectedTileId => {
+          const tiles = this.props.board.tiles.filter(tile => {
+            return tile.id === selectedTileId;
+          })[0];
+
+          return tiles;
+        })
+      : [];
+  };
+
   render() {
     const { navHistory, board, focusTile } = this.props;
 
@@ -1375,7 +1570,7 @@ export class BoardContainer extends Component {
           onLockNotify={this.handleLockNotify}
           onScannerActive={this.handleScannerStrategyNotification}
           onRequestPreviousBoard={this.onRequestPreviousBoard.bind(this)}
-          onRequestToRootBoard={this.props.toRootBoard}
+          onRequestToRootBoard={this.onRequestToRootBoard.bind(this)}
           onSelectClick={this.handleSelectClick}
           onTileClick={this.handleTileClick}
           onBoardTypeChange={this.handleBoardTypeChange}
@@ -1393,6 +1588,13 @@ export class BoardContainer extends Component {
           onTileDrop={this.handleTileDrop}
           onLayoutChange={this.handleLayoutChange}
           disableTour={this.props.disableTour}
+          onCopyTiles={this.handleCopyTiles}
+          onPasteTiles={this.handlePasteTiles}
+          copiedTiles={this.state.copiedTiles}
+          setIsScroll={this.setIsScroll}
+          isScroll={this.state.isScroll}
+          totalRows={this.state.totalRows}
+          ref={this.boardRef}
         />
         <Dialog
           open={!!this.state.copyPublicBoard}
@@ -1456,7 +1658,13 @@ export class BoardContainer extends Component {
           onClose={this.handleTileEditorCancel}
           onEditSubmit={this.handleEditTileEditorSubmit}
           onAddSubmit={this.handleAddTileEditorSubmit}
-          boards={this.props.boards}
+          boards={this.props.boards.filter(
+            board =>
+              board !== null &&
+              board.id !== null &&
+              this.props.communicator.boards.includes(board.id)
+          )}
+          userData={this.props.userData}
         />
       </Fragment>
     );
@@ -1476,23 +1684,17 @@ const mapStateToProps = ({
     communicator => communicator.id === activeCommunicatorId
   );
   const activeBoardId = board.activeBoardId;
-  const currentVoice = speech.voices.find(
-    v => v.voiceURI === speech.options.voiceURI
-  );
   const emptyVoiceAlert =
     speech.voices.length > 0 && speech.options.voiceURI !== EMPTY_VOICES
       ? false
       : true;
-  const offlineVoiceAlert =
-    !isConnected &&
-    speech.voices.length &&
-    currentVoice &&
-    currentVoice.voiceSource === 'cloud';
+  const offlineVoiceAlert = !isConnected && speech.options.isCloud;
   return {
     communicator: currentCommunicator,
     board: board.boards.find(board => board.id === activeBoardId),
     boards: board.boards,
     output: board.output,
+    isLiveMode: board.isLiveMode,
     scannerSettings: scanner,
     navHistory: board.navHistory,
     displaySettings,
@@ -1534,7 +1736,9 @@ const mapDispatchToProps = {
   updateApiObjectsNoChild,
   getApiObjects,
   downloadImages,
-  disableTour
+  disableTour,
+  createApiBoard,
+  upsertApiBoard
 };
 
 export default connect(
