@@ -5,24 +5,8 @@ import IconButton from '@material-ui/core/IconButton';
 import BackspaceIcon from '@material-ui/icons/Backspace';
 import { Scannable } from 'react-scannable';
 
-const styles = {
-  button: {
-    alignSelf: 'center',
-    height: '64px',
-    width: '64px'
-  },
-  icon: {
-    height: '32px',
-    width: '32px'
-  }
-};
-
 export class BackspaceButton extends Component {
   static propTypes = {
-    /**
-     * @ignore
-     */
-    classes: PropTypes.object,
     /**
      * @ignore
      */
@@ -32,23 +16,29 @@ export class BackspaceButton extends Component {
   };
 
   render() {
-    const { classes, theme, hidden, ...other } = this.props;
+    const { theme, hidden, increaseOutputButtons, ...other } = this.props;
 
     const backspaceIconStyle =
       theme.direction === 'ltr' ? null : { transform: 'scaleX(-1)' };
-
     return (
       <Scannable disabled={hidden}>
         <IconButton
           aria-label="Backspace"
-          className={classes.button}
+          className={
+            increaseOutputButtons ? 'Output__button__lg' : 'Output__button__sm'
+          }
           {...other}
         >
-          <BackspaceIcon className={classes.icon} style={backspaceIconStyle} />
+          <BackspaceIcon
+            className={
+              increaseOutputButtons ? 'Output__icon__lg' : 'Output__icon__sm'
+            }
+            style={backspaceIconStyle}
+          />
         </IconButton>
       </Scannable>
     );
   }
 }
 
-export default withStyles(styles, { withTheme: true })(BackspaceButton);
+export default withStyles(null, { withTheme: true })(BackspaceButton);
