@@ -1,42 +1,38 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { withStyles } from '@material-ui/core/styles';
 import IconButton from '@material-ui/core/IconButton';
 import ClearIcon from '@material-ui/icons/Clear';
 import { Scannable } from 'react-scannable';
-
-const styles = {
-  button: {
-    alignSelf: 'center',
-    height: '64px',
-    width: '64px'
-  },
-  icon: {
-    height: '32px',
-    width: '32px'
-  }
-};
 
 export class ClearButton extends Component {
   static propTypes = {
     /**
      * @ignore
      */
-    classes: PropTypes.object,
     hidden: PropTypes.bool
   };
 
   render() {
-    const { classes, hidden, ...other } = this.props;
+    const { hidden, increaseOutputButtons, ...other } = this.props;
 
     return (
       <Scannable disabled={hidden}>
-        <IconButton aria-label="Clear" className={classes.button} {...other}>
-          <ClearIcon className={classes.icon} />
+        <IconButton
+          aria-label="Clear"
+          className={
+            increaseOutputButtons ? 'Output__button__lg' : 'Output__button__sm'
+          }
+          {...other}
+        >
+          <ClearIcon
+            className={
+              increaseOutputButtons ? 'Output__icon__lg' : 'Output__icon__sm'
+            }
+          />
         </IconButton>
       </Scannable>
     );
   }
 }
 
-export default withStyles(styles)(ClearButton);
+export default ClearButton;
