@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 
 import ClearIcon from '@material-ui/icons/Clear';
 import IconButton from '@material-ui/core/IconButton';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import Switch from '@material-ui/core/Switch';
 
 import Symbol from '../../Symbol';
 import BackspaceButton from './BackspaceButton';
@@ -60,9 +62,13 @@ class SymbolOutput extends PureComponent {
       getPhraseToShare,
       onCopyClick,
       onRemoveClick,
+      onSwitchLiveMode,
+      onWriteSymbol,
       symbols,
       navigationSettings,
       phrase,
+      isLiveMode,
+      increaseOutputButtons,
       ...other
     } = this.props;
 
@@ -85,13 +91,23 @@ class SymbolOutput extends PureComponent {
     return (
       <div className="SymbolOutput">
         <Scroll {...other}>
-          {symbols.map(({ image, label }, index) => (
-            <div className="SymbolOutput__value" key={index}>
+          {symbols.map(({ image, label, type }, index) => (
+            <div
+              className={
+                type === 'live'
+                  ? 'LiveSymbolOutput__value'
+                  : 'SymbolOutput__value'
+              }
+              key={index}
+            >
               <Symbol
                 className="SymbolOutput__symbol"
                 image={image}
                 label={label}
+                type={type}
                 labelpos="Below"
+                onWrite={onWriteSymbol(index)}
+                intl={intl}
               />
               <div className="SymbolOutput__value__IconButton">
                 <IconButton
@@ -119,22 +135,49 @@ class SymbolOutput extends PureComponent {
             phrase={this.props.phrase}
             style={copyButtonStyle}
             hidden={!symbols.length}
+            increaseOutputButtons={increaseOutputButtons}
           />
         )}
-        <ClearButton
-          color="inherit"
-          onClick={onClearClick}
-          style={clearButtonStyle}
-          hidden={!symbols.length}
-        />
         {!navigationSettings.removeOutputActive && (
           <BackspaceButton
             color="inherit"
             onClick={onBackspaceClick}
             style={backspaceButtonStyle}
             hidden={navigationSettings.removeOutputActive}
+            increaseOutputButtons={increaseOutputButtons}
           />
         )}
+        <div
+          className={
+            increaseOutputButtons
+              ? 'SymbolOutput__right__btns__lg'
+              : 'SymbolOutput__right__btns'
+          }
+        >
+          {navigationSettings.liveMode && (
+            <FormControlLabel
+              value="bottom"
+              className={increaseOutputButtons ? 'Live__switch_lg' : null}
+              control={
+                <Switch
+                  size="small"
+                  checked={isLiveMode}
+                  color="primary"
+                  onChange={onSwitchLiveMode}
+                />
+              }
+              label={intl.formatMessage(messages.live)}
+              labelPlacement="bottom"
+            />
+          )}
+          <ClearButton
+            color="inherit"
+            onClick={onClearClick}
+            style={clearButtonStyle}
+            hidden={!symbols.length}
+            increaseOutputButtons={increaseOutputButtons}
+          />
+        </div>
       </div>
     );
   }
