@@ -2,9 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { isCordova } from '../../../cordova-util';
+import OutlinedInput from '@material-ui/core/OutlinedInput';
+import messages from '../Board.messages';
 
 import { LABEL_POSITION_BELOW } from '../../Settings/Display/Display.constants';
 import './Symbol.css';
+import { Typography } from '@material-ui/core';
 
 const propTypes = {
   /**
@@ -15,33 +18,68 @@ const propTypes = {
    * Label to display
    */
   label: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
-  labelpos: PropTypes.string
+  labelpos: PropTypes.string,
+  type: PropTypes.string,
+  onWrite: PropTypes.func,
+  intl: PropTypes.object
 };
 
 function Symbol(props) {
-  const { className, label, labelpos, ...other } = props;
+  const { className, label, labelpos, type, onWrite, intl, ...other } = props;
 
   // Cordova path cannot be absolute
   const image =
     isCordova() && props.image && props.image.search('/') === 0
       ? `.${props.image}`
-      : '/cboard-translate' + props.image;
+      : props.image;
 
   const symbolClassName = classNames('Symbol', className);
 
+  const handleKeyPress = event => {
+    if (event.key === 'Enter') {
+      event.preventDefault(); //prevent new line in next textArea
+      return;
+    }
+  };
+
   return (
     <div className={symbolClassName} {...other}>
-      {props.labelpos === 'Above' && props.labelpos !== 'Hidden' && (
-        <div className="Symbol__label">{label}</div>
+      {props.type === 'live' && (
+        <OutlinedInput
+          id="outlined-live-input"
+          margin="none"
+          color="primary"
+          variant="filled"
+          placeholder={intl.formatMessage(messages.writeAndSay)}
+          autoFocus={true}
+          multiline
+          rows={5}
+          value={label}
+          onChange={onWrite}
+          fullWidth={true}
+          onKeyPress={handleKeyPress}
+          style={{
+            padding: '0.5em 0.8em 0.5em 0.8em',
+            height: '100%'
+          }}
+          className={'liveInput'}
+        />
       )}
+      {props.type !== 'live' &&
+        props.labelpos === 'Above' &&
+        props.labelpos !== 'Hidden' && (
+          <Typography className="Symbol__label">{label}</Typography>
+        )}
       {image && (
         <div className="Symbol__image-container">
           <img className="Symbol__image" src={image} alt="" />
         </div>
       )}
-      {props.labelpos === 'Below' && props.labelpos !== 'Hidden' && (
-        <div className="Symbol__label">{label}</div>
-      )}
+      {props.type !== 'live' &&
+        props.labelpos === 'Below' &&
+        props.labelpos !== 'Hidden' && (
+          <Typography className="Symbol__label">{label}</Typography>
+        )}
     </div>
   );
 }
