@@ -30,6 +30,12 @@ import IconButton from '../../UI/IconButton';
 import ColorSelect from '../../UI/ColorSelect';
 import VoiceRecorder from '../../VoiceRecorder';
 import './TileEditor.css';
+import EditIcon from '@material-ui/icons/Edit';
+import ImageEditor from '../ImageEditor';
+
+import API from '../../../api';
+import { isAndroid, writeCvaFile } from '../../../cordova-util';
+import { convertImageUrlToCatchable } from '../../../helpers';
 
 export class TileEditor extends Component {
   static propTypes = {
@@ -57,11 +63,13 @@ export class TileEditor extends Component {
      * Callback fired when submitting a new board tile
      */
     onAddSubmit: PropTypes.func.isRequired,
-    boards: PropTypes.array
+    boards: PropTypes.array,
+    userData: PropTypes.object
   };
 
   static defaultProps = {
-    editingTiles: []
+    editingTiles: [],
+    openImageEditor: false
   };
 
   constructor(props) {
@@ -89,15 +97,34 @@ export class TileEditor extends Component {
       activeStep: 0,
       editingTiles: props.editingTiles,
       isSymbolSearchOpen: false,
+      autoFill: '',
       selectedBackgroundColor: '',
       tile: this.defaultTile,
-      linkedBoard: ''
+      linkedBoard: '',
+      imageUploadedData: [],
+      isEditImageBtnActive: false
+    };
+
+    this.defaultimageUploadedData = {
+      isUploaded: false,
+      fileName: '',
+      blobHQ: null,
+      blob: null
     };
   }
 
   UNSAFE_componentWillReceiveProps(props) {
     this.updateTileProperty('id', shortid.generate()); // todo not here
     this.setState({ editingTiles: props.editingTiles });
+  }
+  componentDidUpdate(prevProps) {
+    if (
+      this.props.open !== prevProps.open &&
+      this.props.open &&
+      this.editingTile()
+    ) {
+      this.setLinkedBoard();
+    }
   }
 
   editingTile() {
@@ -135,51 +162,169 @@ export class TileEditor extends Component {
     }
   }
 
-  handleSubmit = () => {
+  handleSubmit = async () => {
     const { onEditSubmit, onAddSubmit } = this.props;
-
-    this.setState({
-      activeStep: 0,
-      selectedBackgroundColor: '',
-      tile: this.defaultTile
-    });
-
     if (this.editingTile()) {
-      onEditSubmit(this.state.editingTiles);
+      const { imageUploadedData } = this.state;
+      if (imageUploadedData.length) {
+        let tilesToAdd = JSON.parse(JSON.stringify(this.state.editingTiles));
+        await Promise.all(
+          imageUploadedData.map(async (obj, index) => {
+            if (obj.isUploaded) {
+              tilesToAdd[index].image = await this.updateTileImgURL(
+                obj.blob,
+                obj.fileName
+              );
+            }
+          })
+        );
+        onEditSubmit(tilesToAdd);
+      } else {
+        onEditSubmit(this.state.editingTiles);
+      }
     } else {
       const tileToAdd = this.state.tile;
-      const selectedBackgroundColor = this.state.selectedBackgroundColor;
+      const imageUploadedData = this.state.imageUploadedData[
+        this.state.activeStep
+      ];
+      if (imageUploadedData && imageUploadedData.isUploaded) {
+        tileToAdd.image = await this.updateTileImgURL(
+          imageUploadedData.blob,
+          imageUploadedData.fileName
+        );
+      }
 
+      const selectedBackgroundColor = this.state.selectedBackgroundColor;
       if (selectedBackgroundColor) {
         tileToAdd.backgroundColor = selectedBackgroundColor;
       }
       onAddSubmit(tileToAdd);
     }
-  };
-
-  handleCancel = () => {
-    const { onClose } = this.props;
 
     this.setState({
       activeStep: 0,
       selectedBackgroundColor: '',
-      tile: this.defaultTile
+      tile: this.defaultTile,
+      imageUploadedData: [],
+      isEditImageBtnActive: false,
+      linkedBoard: ''
+    });
+  };
+
+  updateTileImgURL = async (blob, fileName) => {
+    const { userData } = this.props;
+    const user = userData.email ? userData : null;
+    if (user) {
+      // this.setState({
+      //   loading: true
+      // });
+      try {
+        const imageUrl = await API.uploadFile(blob, fileName);
+        // console.log('imagen guardada en servidor', imageUrl);
+        return convertImageUrlToCatchable(imageUrl) || imageUrl;
+      } catch (error) {
+        //console.log('imagen no guardad en servidor');
+        return await this.blobToBase64(blob);
+      }
+      // } finally {
+      //   this.setState({
+      //     loading: false
+      //   });
+    } else {
+      if (isAndroid()) {
+        const filePath = '/Android/data/com.unicef.cboard/files/' + fileName;
+        const fEntry = await writeCvaFile(filePath, blob);
+        return fEntry.nativeURL;
+      } else {
+        return await this.blobToBase64(blob);
+      }
+    }
+  };
+
+  blobToBase64 = async blob => {
+    return new Promise(resolve => {
+      const reader = new FileReader();
+      reader.onload = () => {
+        resolve(reader.result);
+      };
+      reader.readAsDataURL(blob);
+    });
+  };
+
+  handleCancel = () => {
+    const { onClose } = this.props;
+    this.setState({
+      activeStep: 0,
+      selectedBackgroundColor: '',
+      tile: this.defaultTile,
+      imageUploadedData: [],
+      isEditImageBtnActive: false,
+      linkedBoard: ''
     });
     onClose();
   };
 
-  handleInputImageChange = image => {
+  createimageUploadedDataArray() {
+    if (this.editingTile()) {
+      let imageUploadedDataArray = new Array(this.state.editingTiles.length);
+      imageUploadedDataArray.fill(this.defaultimageUploadedData);
+      this.setState({ imageUploadedData: imageUploadedDataArray });
+    } else {
+      this.setState({
+        imageUploadedData: new Array(this.defaultimageUploadedData)
+      });
+    }
+  }
+
+  handleInputImageChange = (blob, fileName, blobHQ) => {
+    if (!this.state.imageUploadedData.length) {
+      this.createimageUploadedDataArray();
+    }
+    this.setimageUploadedData(true, fileName, blobHQ, blob);
+    this.setState({ isEditImageBtnActive: true });
+    const image = URL.createObjectURL(blob);
     this.updateTileProperty('image', image);
+  };
+
+  setimageUploadedData = (isUploaded, fileName, blobHQ = null, blob = null) => {
+    const { activeStep } = this.state;
+    let imageUploadedData = this.state.imageUploadedData.map((item, indx) => {
+      if (indx === activeStep) {
+        return {
+          ...item,
+          isUploaded: isUploaded,
+          fileName: fileName,
+          blobHQ: blobHQ,
+          blob: blob
+        };
+      } else {
+        return item;
+      }
+    });
+    this.setState({ imageUploadedData: imageUploadedData });
   };
 
   handleSymbolSearchChange = ({ image, labelKey, label }) => {
-    this.updateTileProperty('labelKey', labelKey);
-    this.updateTileProperty('label', label);
-    this.updateTileProperty('image', image);
+    return new Promise(resolve => {
+      this.updateTileProperty('labelKey', labelKey);
+      this.updateTileProperty('label', label);
+      this.updateTileProperty('image', image);
+      if (this.state.imageUploadedData.length) {
+        this.setimageUploadedData(false, '');
+      }
+      resolve();
+    });
   };
 
   handleSymbolSearchClose = event => {
+    const { imageUploadedData } = this.state;
     this.setState({ isSymbolSearchOpen: false });
+    if (
+      imageUploadedData.length &&
+      imageUploadedData[this.state.activeStep].isUploaded
+    ) {
+      this.setState({ isEditImageBtnActive: true });
+    }
   };
 
   handleLabelChange = event => {
@@ -216,17 +361,24 @@ export class TileEditor extends Component {
   };
 
   handleBack = event => {
-    this.setState({ activeStep: this.state.activeStep - 1 });
-    this.setState({ selectedBackgroundColor: '', linkedBoard: '' });
+    this.setState({ activeStep: this.state.activeStep - 1 }, () => {
+      this.setLinkedBoard();
+    });
+    this.setState({ selectedBackgroundColor: '' });
+    this.setState({ isEditImageBtnActive: false });
   };
 
-  handleNext = event => {
-    this.setState({ activeStep: this.state.activeStep + 1 });
-    this.setState({ selectedBackgroundColor: '', linkedBoard: '' });
+  handleNext = async event => {
+    this.setState({ activeStep: this.state.activeStep + 1 }, () => {
+      this.setLinkedBoard();
+    });
+    this.setState({ selectedBackgroundColor: '' });
+    this.setState({ isEditImageBtnActive: false });
   };
 
-  handleSearchClick = event => {
-    this.setState({ isSymbolSearchOpen: true });
+  handleSearchClick = (event, currentLabel) => {
+    this.setState({ isSymbolSearchOpen: true, autoFill: currentLabel || '' });
+    this.setState({ isEditImageBtnActive: false });
   };
 
   handleColorChange = event => {
@@ -259,24 +411,50 @@ export class TileEditor extends Component {
       this.updateTileProperty('loadBoard', board.id);
     } else {
       this.updateTileProperty('linkedBoard', false);
+      this.updateTileProperty('loadBoard', shortid.generate());
     }
+  };
+
+  handleOnClickImageEditor = () => {
+    this.setState({ openImageEditor: true });
+  };
+  onImageEditorClose = () => {
+    this.setState({ openImageEditor: false });
+  };
+  onImageEditorDone = blob => {
+    this.setState(prevState => {
+      const newArray = [...prevState.imageUploadedData];
+      newArray[this.state.activeStep].blob = blob;
+      return { imageUploadedData: newArray };
+    });
+    const image = URL.createObjectURL(blob);
+    this.updateTileProperty('image', image);
+  };
+
+  setLinkedBoard = () => {
+    const loadBoard =
+      this.currentTileProp('linkedBoard') || this.editingTile()
+        ? this.currentTileProp('loadBoard')
+        : null;
+    const linkedBoard =
+      this.props.boards.find(board => board.id === loadBoard) || 'none';
+    this.setState({ linkedBoard: linkedBoard });
   };
 
   render() {
     const { open, intl, boards } = this.props;
-
     const currentLabel = this.currentTileProp('labelKey')
       ? intl.formatMessage({ id: this.currentTileProp('labelKey') })
       : this.currentTileProp('label');
-
     const buttons = (
       <IconButton
         label={intl.formatMessage(messages.symbolSearch)}
-        onClick={this.handleSearchClick}
+        onClick={e => this.handleSearchClick(e, currentLabel)}
       >
         <SearchIcon />
       </IconButton>
     );
+
     const selectBoardElement = (
       <div>
         <FormControl fullWidth>
@@ -290,9 +468,11 @@ export class TileEditor extends Component {
             value={this.state.linkedBoard}
             onChange={this.handleBoardsChange}
           >
-            <MenuItem value="none">
-              <em>{intl.formatMessage(messages.none)}</em>
-            </MenuItem>
+            {!this.editingTile() && (
+              <MenuItem value="none">
+                <em>{intl.formatMessage(messages.none)}</em>
+              </MenuItem>
+            )}
             {boards.map(
               board =>
                 !board.hidden && (
@@ -343,11 +523,34 @@ export class TileEditor extends Component {
                         <Symbol image={tileInView.image} label={currentLabel} />
                       </Tile>
                     </div>
+                    {this.state.isEditImageBtnActive && (
+                      <React.Fragment>
+                        <ImageEditor
+                          intl={intl}
+                          open={this.state.openImageEditor}
+                          onImageEditorClose={this.onImageEditorClose}
+                          onImageEditorDone={this.onImageEditorDone}
+                          image={URL.createObjectURL(
+                            this.state.imageUploadedData[this.state.activeStep]
+                              .blobHQ
+                          )}
+                        />
+                        <Button
+                          variant="contained"
+                          color="secondary"
+                          startIcon={<EditIcon />}
+                          onClick={this.handleOnClickImageEditor}
+                          style={{ marginBottom: '6px' }}
+                        >
+                          {intl.formatMessage(messages.editImage)}
+                        </Button>
+                      </React.Fragment>
+                    )}
                     <Button
                       variant="contained"
                       color="primary"
                       startIcon={<SearchIcon />}
-                      onClick={this.handleSearchClick}
+                      onClick={e => this.handleSearchClick(e, currentLabel)}
                     >
                       {intl.formatMessage(messages.symbols)}
                     </Button>
@@ -378,11 +581,6 @@ export class TileEditor extends Component {
                       onChange={this.handleVocalizationChange}
                       fullWidth
                     />
-                    <div>
-                      {this.editingTile() &&
-                        tileInView.loadBoard &&
-                        selectBoardElement}
-                    </div>
                     {!this.editingTile() && (
                       <div className="TileEditor__radiogroup">
                         <FormControl fullWidth>
@@ -478,6 +676,7 @@ export class TileEditor extends Component {
 
           <SymbolSearch
             open={this.state.isSymbolSearchOpen}
+            autoFill={this.state.autoFill}
             onChange={this.handleSymbolSearchChange}
             onClose={this.handleSymbolSearchClose}
           />
