@@ -135,7 +135,7 @@ describe('actions', () => {
     };
     expect(actions.toRootBoard()).toEqual(expectedAction);
   });
-  
+
   it('should create an action to REPLACE_ME', () => {
     const boardId = '123';
     const expectedAction = {
@@ -303,6 +303,13 @@ describe('actions', () => {
     store
       .dispatch(actions.createApiBoard(mockBoard, '12345678901234567'))
       .then(data => {
+        const actions = store.getActions();
+        const dataResp = {
+          board: mockBoard,
+          boardId: '12345678901234567',
+          type: 'cboard/Board/CREATE_API_BOARD_SUCCESS'
+        };
+        expect(actions[1]).toEqual(dataResp);
         expect(data).toEqual(mockBoard);
       })
       .catch(e => {
@@ -313,11 +320,16 @@ describe('actions', () => {
     const store = mockStore(initialState);
     store
       .dispatch(actions.createApiBoard({ error: 'error' }, '12345678901234567'))
-      .then(data => {
-        expect(data).toEqual(mockBoard);
+      .then(() => {
+        throw new Error('An error was expected');
       })
       .catch(e => {
-        throw new Error(e.message);
+        const actions = store.getActions();
+        const dataResp = {
+          message: '[object Object]',
+          type: 'cboard/Board/CREATE_API_BOARD_FAILURE'
+        };
+        expect(actions[1]).toEqual(dataResp);
       });
   });
   it('check updateApiBoard', () => {
@@ -335,11 +347,16 @@ describe('actions', () => {
     const store = mockStore(initialState);
     store
       .dispatch(actions.updateApiBoard({ error: 'error' }))
-      .then(data => {
-        expect(data).toEqual(mockBoard);
+      .then(() => {
+        throw new Error('An error was expected');
       })
       .catch(e => {
-        throw new Error(e.message);
+        const actions = store.getActions();
+        const dataResp = {
+          message: '[object Object]',
+          type: 'cboard/Board/UPDATE_API_BOARD_FAILURE'
+        };
+        expect(actions[1]).toEqual(dataResp);
       });
   });
   it('check deleteApiBoard', () => {
@@ -357,11 +374,16 @@ describe('actions', () => {
     const store = mockStore(initialState);
     store
       .dispatch(actions.deleteApiBoard('error'))
-      .then(data => {
-        expect(data).toEqual(mockBoard);
+      .then(() => {
+        throw new Error('An error was expected');
       })
       .catch(e => {
-        throw new Error(e.message);
+        const actions = store.getActions();
+        const dataResp = {
+          message: '[object Object]',
+          type: 'cboard/Board/DELETE_API_BOARD_FAILURE'
+        };
+        expect(actions[1]).toEqual(dataResp);
       });
   });
   it('check updateApiObjectsNoChild', () => {
