@@ -9,28 +9,49 @@ const NavigationButtons = ({
   active,
   navHistory,
   previousBoard,
-  toRootBoard
+  toRootBoard,
+  isLocked,
+  isNavigationButtonsOnTheSide
 }) => {
   if (!active) {
     return null;
   }
+  const classPreviousBoardButton = isNavigationButtonsOnTheSide
+    ? `SideNavigationButton SideButtonPreviousBoard ${
+        navHistory.length > 2 ? '' : 'disable'
+      }`
+    : `NavigationButton left`;
+
+  const classToRootBoardButton = isNavigationButtonsOnTheSide
+    ? `SideNavigationButton SideButtonToRootBoard ${
+        navHistory.length > 1 ? '' : 'disable'
+      }`
+    : 'NavigationButton right';
 
   return (
     <React.Fragment>
-      {navHistory.length > 2 && (
-        <div className="NavigationButton left">
-          <button onClick={toRootBoard}>
-            <FirstPageIcon />
-          </button>
-        </div>
-      )}
-      {navHistory.length > 1 && (
-        <div className="NavigationButton right">
-          <button onClick={previousBoard}>
-            <ChevronLeftIcon />
-          </button>
-        </div>
-      )}
+      <div
+        className={
+          isNavigationButtonsOnTheSide
+            ? `SideNavigationButtonsContainer ${!isLocked ? 'moveDown' : ''}`
+            : ''
+        }
+      >
+        {(navHistory.length > 2 || isNavigationButtonsOnTheSide) && (
+          <div className={classPreviousBoardButton}>
+            <button onClick={toRootBoard}>
+              <FirstPageIcon />
+            </button>
+          </div>
+        )}
+        {(navHistory.length > 1 || isNavigationButtonsOnTheSide) && (
+          <div className={classToRootBoardButton}>
+            <button onClick={previousBoard}>
+              <ChevronLeftIcon />
+            </button>
+          </div>
+        )}
+      </div>
     </React.Fragment>
   );
 };
@@ -38,7 +59,9 @@ const NavigationButtons = ({
 NavigationButtons.props = {
   navHistory: PropTypes.arrayOf(PropTypes.string),
   previousBoard: PropTypes.func,
-  toRootBoard: PropTypes.func
+  toRootBoard: PropTypes.func,
+  isLocked: PropTypes.bool,
+  isNavigationButtonsOnTheSide: PropTypes.bool
 };
 
 export default NavigationButtons;
