@@ -10,6 +10,10 @@ import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
 import Divider from '@material-ui/core/Divider';
 import FullScreenDialog from '../../UI/FullScreenDialog';
 import messages from './Navigation.messages';
+import MenuItem from '@material-ui/core/MenuItem';
+import FormControl from '@material-ui/core/FormControl';
+import Select from '@material-ui/core/Select';
+import { NAVIGATION_BUTTONS_STYLES } from './Navigation.constants';
 
 import './Navigation.css';
 import ResetToursItem from '../../UI/ResetToursItem';
@@ -20,7 +24,9 @@ const propTypes = {
    */
   onClose: PropTypes.func,
   updateNavigationSettings: PropTypes.func.isRequired,
-  navigationSettings: PropTypes.object.isRequired
+  navigationSettings: PropTypes.object.isRequired,
+  isLiveMode: PropTypes.bool,
+  changeLiveMode: PropTypes.func.isRequired
 };
 
 class Navigation extends React.Component {
@@ -35,6 +41,12 @@ class Navigation extends React.Component {
   toggleCABackButton = () => {
     this.setState({
       caBackButtonActive: !this.state.caBackButtonActive
+    });
+  };
+
+  toggleCAScrollButton = () => {
+    this.setState({
+      bigScrollButtonsActive: !this.state.bigScrollButtonsActive
     });
   };
 
@@ -62,9 +74,58 @@ class Navigation extends React.Component {
     });
   };
 
+  toggleLiveMode = () => {
+    this.setState({
+      liveMode: !this.state.liveMode
+    });
+  };
+
   onSubmit = () => {
+    const { isLiveMode, changeLiveMode } = this.props;
+    if (!this.state.liveMode && isLiveMode) {
+      changeLiveMode();
+    }
     this.props.updateNavigationSettings(this.state);
   };
+
+  onNavigationSettingsChange(navigationSetting, event) {
+    const {
+      target: { value }
+    } = event;
+    this.setState({ [navigationSetting]: value });
+  }
+
+  renderNavigationButtonsLocationSelect() {
+    const name = 'navigationButtonsStyle';
+    const actualButtonsStyle = NAVIGATION_BUTTONS_STYLES.filter(
+      style => style.value === this.state[name]
+    )[0];
+
+    return (
+      <FormControl>
+        <Select
+          aria-label={name}
+          id={name}
+          name={name}
+          value={
+            actualButtonsStyle?.value || NAVIGATION_BUTTONS_STYLES[0].value
+          }
+          onChange={e => this.onNavigationSettingsChange(name, e)}
+          disabled={
+            !(
+              this.state.bigScrollButtonsActive || this.state.caBackButtonActive
+            )
+          }
+        >
+          {NAVIGATION_BUTTONS_STYLES.map(style => (
+            <MenuItem key={style?.value} value={style?.value}>
+              {style?.name}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+    );
+  }
 
   render() {
     const { onClose } = this.props;
@@ -91,6 +152,48 @@ class Navigation extends React.Component {
                     value="active"
                     color="secondary"
                   />
+                </ListItemSecondaryAction>
+              </ListItem>
+              <Divider />
+              <ListItem>
+                <ListItemText
+                  className="Navigation__ListItemText"
+                  primary={<FormattedMessage {...messages.bigScroll} />}
+                  secondary={
+                    <FormattedMessage {...messages.bigScrollSecondary} />
+                  }
+                />
+                <ListItemSecondaryAction>
+                  <Switch
+                    checked={this.state.bigScrollButtonsActive}
+                    onChange={this.toggleCAScrollButton}
+                    value="active"
+                    color="secondary"
+                  />
+                </ListItemSecondaryAction>
+              </ListItem>
+              <Divider />
+              <ListItem
+                disabled={
+                  !(
+                    this.state.bigScrollButtonsActive ||
+                    this.state.caBackButtonActive
+                  )
+                }
+              >
+                <ListItemText
+                  className="Display__ListItemText"
+                  primary={
+                    <FormattedMessage {...messages.navigationButtonsStyle} />
+                  }
+                  secondary={
+                    <FormattedMessage
+                      {...messages.navigationButtonsStyleSecondary}
+                    />
+                  }
+                />
+                <ListItemSecondaryAction className="Display__Options">
+                  {this.renderNavigationButtonsLocationSelect()}
                 </ListItemSecondaryAction>
               </ListItem>
               <Divider />
@@ -168,6 +271,24 @@ class Navigation extends React.Component {
               </ListItem>
               <Divider />
               <ResetToursItem />
+              <Divider />
+              <ListItem>
+                <ListItemText
+                  className="Display__ListItemText"
+                  primary={<FormattedMessage {...messages.showLiveMode} />}
+                  secondary={
+                    <FormattedMessage {...messages.showLiveModeSecondary} />
+                  }
+                />
+                <ListItemSecondaryAction>
+                  <Switch
+                    checked={this.state.liveMode || false}
+                    onChange={this.toggleLiveMode}
+                    value="active"
+                    color="secondary"
+                  />
+                </ListItemSecondaryAction>
+              </ListItem>
             </List>
           </Paper>
         </FullScreenDialog>
