@@ -11,7 +11,8 @@ export default defineMessages({
   },
   enableSecondary: {
     id: 'cboard.components.Settings.Navigation.enableSecondary',
-    defaultMessage: 'Shows big back buttons on top of the boards'
+    defaultMessage:
+      'Shows big back buttons. Select the desired style in navigation buttons style section.'
   },
   quickUnlock: {
     id: 'cboard.components.Settings.Navigation.quickUnlock',
@@ -45,5 +46,32 @@ export default defineMessages({
   vocalizeFoldersSecondary: {
     id: 'cboard.components.Settings.Navigation.vocalizeFoldersSecondary',
     defaultMessage: "Reads a folder's name out loud when clicked"
+  },
+  showLiveMode: {
+    id: 'cboard.components.Settings.Navigation.showLiveMode',
+    defaultMessage: 'Use the Live Mode'
+  },
+  showLiveModeSecondary: {
+    id: 'cboard.components.Settings.Navigation.showLiveModeSecondary',
+    defaultMessage:
+      'Live mode allows you to write text directly into the output bar and quickly play the sound. It is intended for users that can write.'
+  },
+  bigScroll: {
+    id: 'cboard.components.Settings.Navigation.bigScroll',
+    defaultMessage: 'Enable big scroll buttons'
+  },
+  bigScrollSecondary: {
+    id: 'cboard.components.Settings.Navigation.bigScrollSecondary',
+    defaultMessage:
+      'Shows big scroll buttons. Select the desired style in navigation buttons style section'
+  },
+  navigationButtonsStyle: {
+    id: 'cboard.components.Settings.Navigation.navigationButtonsStyle',
+    defaultMessage: 'Navigation buttons style'
+  },
+  navigationButtonsStyleSecondary: {
+    id: 'cboard.components.Settings.Navigation.navigationButtonsStyleSecondary',
+    defaultMessage:
+      'Select the style of the context-aware back and big scroll buttons. On the sides is recommended for eye trackers'
   }
 });
