@@ -50,14 +50,22 @@ export class PeopleContainer extends PureComponent {
   handleLogout = () => {
     if (isAndroid()) {
       window.plugins.googleplus.disconnect(function(msg) {
-        console.log('disconnect msg' + msg);
+        console.log('disconnect google msg' + msg);
       });
+      window.facebookConnectPlugin.logout(
+        function(msg) {
+          console.log('disconnect facebook msg' + msg);
+        },
+        function(msg) {
+          console.log('error facebook disconnect msg' + msg);
+        }
+      );
     }
     this.props.logout();
   };
 
   render() {
-    const { history } = this.props;
+    const { history, location } = this.props;
 
     return (
       <People
@@ -67,6 +75,7 @@ export class PeopleContainer extends PureComponent {
         name={this.state.name}
         email={this.state.email}
         birthdate={this.state.birthdate}
+        location={location}
         onChangePeople={this.handleChange}
         onSubmitPeople={this.handleSubmit}
       />
@@ -74,10 +83,24 @@ export class PeopleContainer extends PureComponent {
   }
 }
 
-const mapStateToProps = state => ({
-  isLogged: isLogged(state),
-  user: getUser(state)
-});
+const mapStateToProps = state => {
+  const userIsLogged = isLogged(state);
+  const user = getUser(state);
+  const location = userIsLogged
+    ? {
+        country: user?.location?.country,
+        countryCode: user?.location?.countryCode
+      }
+    : {
+        country: state.app.unloggedUserLocation?.country,
+        countryCode: state.app.unloggedUserLocation?.countryCode
+      };
+  return {
+    isLogged: userIsLogged,
+    user: user,
+    location: location
+  };
+};
 
 const mapDispatchToProps = {
   logout: logout,
