@@ -21,18 +21,29 @@ import LiveHelpIcon from '@material-ui/icons/LiveHelp';
 import messages from './Settings.messages';
 import SettingsSection from './SettingsSection.component';
 import FullScreenDialog from '../UI/FullScreenDialog';
+import Paper from '@material-ui/core/Paper';
 import UserIcon from '../UI/UserIcon';
 import SettingsTour from './SettingsTour.component';
 
-import { isAndroid } from '../../cordova-util';
+import { isCordova, isAndroid } from '../../cordova-util';
 
 import './Settings.css';
+import { CircularProgress } from '@material-ui/core';
+
+import { Adsense } from '@ctrl/react-adsense';
+import {
+  ADSENSE_ON_PRODUCTION,
+  ADTEST_AVAILABLE,
+  ADSENSE_CLIENT,
+  ADD_SLOT_SETTINGS_TOP
+} from '../../constants';
 
 const propTypes = {
   isLogged: PropTypes.bool.isRequired,
   logout: PropTypes.func.isRequired,
   user: PropTypes.object.isRequired,
-  intl: intlShape.isRequired
+  intl: intlShape.isRequired,
+  isDownloadingLang: PropTypes.bool
 };
 
 export class Settings extends PureComponent {
@@ -42,8 +53,16 @@ export class Settings extends PureComponent {
     function handleLogOutClick() {
       if (isAndroid()) {
         window.plugins.googleplus.disconnect(function(msg) {
-          console.log('disconnect msg' + msg);
+          console.log('disconnect google msg' + msg);
         });
+        window.facebookConnectPlugin.logout(
+          function(msg) {
+            console.log('disconnect facebook msg' + msg);
+          },
+          function(msg) {
+            console.log('error facebook disconnect msg' + msg);
+          }
+        );
       }
       logout();
     }
@@ -163,7 +182,8 @@ export class Settings extends PureComponent {
   };
 
   handleGoBack = () => {
-    const { history } = this.props;
+    const { history, isDownloadingLang } = this.props;
+    if (isDownloadingLang) return; //prevent goBack during downloading
     history.replace('/');
   };
 
@@ -173,7 +193,13 @@ export class Settings extends PureComponent {
   };
 
   render() {
-    const { intl, disableTour, isSettingsTourEnabled, location } = this.props;
+    const {
+      intl,
+      disableTour,
+      isSettingsTourEnabled,
+      location,
+      isDownloadingLang
+    } = this.props;
     const isSettingsLocation = location.pathname === '/settings';
     return (
       <FullScreenDialog
@@ -183,7 +209,8 @@ export class Settings extends PureComponent {
         onClose={this.handleGoBack}
         buttons={
           isSettingsLocation &&
-          !isSettingsTourEnabled && (
+          !isSettingsTourEnabled &&
+          !isDownloadingLang && (
             <div className="Settings_EnableTour_Button">
               <IconButton
                 label={intl.formatMessage(messages.enableTour)}
@@ -195,13 +222,42 @@ export class Settings extends PureComponent {
           )
         }
       >
-        {this.getSettingsSections().map(({ subheader, settings }, index) => (
-          <SettingsSection
-            subheader={subheader}
-            settings={settings}
-            key={index}
-          />
-        ))}
+        {!isCordova() && (
+          <Paper className="Settings__section">
+            <Adsense
+              style={{
+                display: 'block',
+                height: '30vh',
+                maxHeight: '198px'
+              }}
+              client={ADSENSE_CLIENT}
+              slot={ADD_SLOT_SETTINGS_TOP}
+              data-adtest={ADSENSE_ON_PRODUCTION ? 'off' : 'on'}
+              format="none"
+              className={
+                ADSENSE_ON_PRODUCTION || ADTEST_AVAILABLE
+                  ? null
+                  : 'adSense__test__marker'
+              }
+            />
+          </Paper>
+        )}
+        {(isDownloadingLang && (
+          <div className="Settings__spinner-container">
+            <CircularProgress
+              size={60}
+              className="Settings__loading-Spinner"
+              thickness={4}
+            />
+          </div>
+        )) ||
+          this.getSettingsSections().map(({ subheader, settings }, index) => (
+            <SettingsSection
+              subheader={subheader}
+              settings={settings}
+              key={index}
+            />
+          ))}
         {isSettingsLocation && isSettingsTourEnabled && (
           <SettingsTour
             intl={intl}
