@@ -21,8 +21,8 @@ export default defineMessages({
     id: 'cboard.components.WelcomeScreen.skipForNow',
     defaultMessage: 'Skip for now'
   },
-  googleLoginErrorAndroid: {
-    id: 'cboard.components.WelcomeScreen.googleLoginErrorAndroid',
+  loginErrorAndroid: {
+    id: 'cboard.components.WelcomeScreen.loginErrorAndroid',
     defaultMessage:
       'Sorry. An error occurred during authenticate process. Please try it again'
   },
