@@ -8,10 +8,6 @@ export const GLOBALSYMBOLS_BASE_PATH_API = 'https://globalsymbols.com/api/v1/';
 export const API_URL = isCordova()
   ? 'https://api.app.cboard.io/'
   : DEV_API_URL || `${window.location.protocol}//api.${host}`;
-export const TAWASOL_BASE_PATH_API =
-  'https://madaportal.org/tawasol/wp-json/custom/v2/dea44ade76/';
-export const TAWASOL_BASE_IMAGE_URL =
-  'https://www.madaportal.org/tawasol/wp-content/uploads/images/';
 export const AZURE_INST_KEY =
   process.env.REACT_APP_AZURE_INST_KEY ||
   '874487ac-304c-4160-b8f3-a221541eab61';
@@ -23,3 +19,15 @@ export const AZURE_VOICES_BASE_PATH_API =
   'https://' +
   AZURE_SPEECH_SERVICE_REGION +
   '.tts.speech.microsoft.com/cognitiveservices/voices/';
+
+// AdSense constants
+
+const NODE_ENV = process.env.NODE_ENV;
+const HOSTNAME = window.location.hostname;
+export const ADSENSE_ON_PRODUCTION =
+  HOSTNAME === 'app.cboard.io' && NODE_ENV === 'production';
+export const ADTEST_AVAILABLE =
+  HOSTNAME === 'app.dev.cboard.io' || HOSTNAME === 'app.qa.cboard.io';
+export const ADSENSE_CLIENT = 'ca-pub-7162313874228987';
+
+export const ADD_SLOT_SETTINGS_TOP = '5250438005';
