@@ -1,4 +1,8 @@
-import { CHANGE_LANG, SET_LANGS } from './LanguageProvider.constants';
+import {
+  CHANGE_LANG,
+  SET_LANGS,
+  SET_DOWNLOADING_LANG
+} from './LanguageProvider.constants';
 import { updateUserData } from '../../components/App/App.actions';
 
 export function changeLang(lang) {
@@ -14,5 +18,12 @@ export function setLangs(langs, localLangs) {
     type: SET_LANGS,
     langs,
     localLangs
+  };
+}
+
+export function setDownloadingLang(downloadingLangData) {
+  return {
+    type: SET_DOWNLOADING_LANG,
+    downloadingLangData
   };
 }
