@@ -24,6 +24,7 @@ import {
   filterLocalLangs
 } from '../../i18n';
 import tts from './tts';
+import { showNotification } from '../../components/Notifications/Notifications.actions';
 
 export function requestVoices() {
   return {
@@ -52,19 +53,26 @@ export function receiveTtsEngine(ttsEngineName) {
 }
 
 export function getTtsEngines() {
-  const ttsEngines = tts.getTtsEngines();
+  const ttsEngines = tts?.getTtsEngines();
   return {
     type: RECEIVE_TTS_ENGINES,
     ttsEngines
   };
 }
 
-export function setTtsEngine(ttsEngineName) {
+export function setTtsEngine(selectedTtsEngineName) {
   return async dispatch => {
     dispatch(requestTtsEngine());
     try {
-      const voices = await tts.setTtsEngine(ttsEngineName);
-      dispatch(receiveTtsEngine(ttsEngineName));
+      const engineAvailable = tts
+        .getTtsEngines()
+        .map(tts => tts.name)
+        .includes(selectedTtsEngineName);
+      const engineName = engineAvailable
+        ? selectedTtsEngineName
+        : tts.getTtsDefaultEngine().name;
+      const voices = await tts.setTtsEngine(engineName);
+      dispatch(receiveTtsEngine(engineName));
       if (!voices.length) {
         throw new Error('TTS engine does not have a language.');
       }
@@ -77,12 +85,13 @@ export function setTtsEngine(ttsEngineName) {
 export function updateLangSpeechStatus(voices) {
   return async (dispatch, getState) => {
     try {
-      // once we got the voices, we calculate what are the supported languages
       const supportedLangs = getSupportedLangs(voices);
+
       if (!supportedLangs.length) {
         throw new Error('TTS engine does not have a supported language.');
       }
       const localLangs = filterLocalLangs(voices);
+
       dispatch(setLangs(supportedLangs, localLangs));
 
       // now we set the actual language based on the state
@@ -126,10 +135,17 @@ export function getTtsDefaultEngine() {
 }
 
 export function changeVoice(voiceURI, lang) {
-  return {
-    type: CHANGE_VOICE,
-    voiceURI,
-    lang
+  return (dispatch, getState) => {
+    const isCloud =
+      getState().speech.voices.find(v => v.voiceURI === voiceURI)
+        ?.voiceSource === 'cloud';
+    if (isCloud) dispatch(showNotification('', 'cloudVoiceIsSeted'));
+    dispatch({
+      type: CHANGE_VOICE,
+      voiceURI,
+      lang,
+      isCloud
+    });
   };
 }
 
@@ -152,6 +168,18 @@ export function getVoices() {
     let voices = [];
     dispatch(requestVoices());
     try {
+      const localizeSerbianVoicesNames = (voiceName, voiceLang) => {
+        if (voiceLang?.startsWith('sr')) {
+          const getNativeNameOfDialect = lang => {
+            if (lang === 'sr-ME') return 'Crnogorski jezik';
+            if (lang === 'sr-SP') return 'Српски језик';
+            if (lang === 'sr-RS') return 'Srpski jezik';
+          };
+          return `${voiceName} - ${getNativeNameOfDialect(voiceLang)}`;
+        }
+        return voiceName;
+      };
+
       const pvoices = await tts.getVoices();
       // some TTS engines do return invalid voices, so we filter them
       const regex = new RegExp('^[a-zA-Z]{2,}-$', 'g');
@@ -176,6 +204,7 @@ export function getVoices() {
           } else if (DisplayName) {
             voice.name = `${DisplayName} (${voice.lang}) - ${Gender}`;
           }
+          voice.name = localizeSerbianVoicesNames(voice.name, voice.lang);
           return voice;
         }
       );
@@ -210,21 +239,46 @@ export function cancelSpeech() {
       type: CANCEL_SPEECH,
       isSpeaking: false
     });
-    tts.cancel();
+    try {
+      tts.cancel();
+    } catch (error) {
+      console.err(error);
+    }
   };
 }
 
 export function speak(text, onend = () => {}) {
   return (dispatch, getState) => {
     const options = getState().speech.options;
+    const setCloudSpeakAlertTimeout = () => {
+      const REASONABLE_TIME_TO_AWAIT = 5000;
+      return setTimeout(() => {
+        dispatch(showNotification('', 'cloudSpeakError'));
+      }, REASONABLE_TIME_TO_AWAIT);
+    };
     dispatch(startSpeech(text));
 
-    tts.speak(text, {
-      ...options,
-      onend: event => {
-        onend();
-        dispatch(endSpeech());
-      }
-    });
+    tts.speak(
+      text,
+      {
+        ...options,
+        onend: event => {
+          onend();
+          dispatch(endSpeech());
+          if (event?.error) dispatch(showNotification('', 'cloudSpeakError'));
+        }
+      },
+      setCloudSpeakAlertTimeout
+    );
+  };
+}
+
+export function setCurrentVoiceSource() {
+  return (dispatch, getState) => {
+    const { isCloud = null, voiceURI, lang } = getState().speech.options;
+    if (isCloud === null && !!voiceURI && !!lang) {
+      dispatch(changeVoice(voiceURI, lang));
+      return;
+    }
   };
 }
