@@ -30,6 +30,7 @@ const initialState = {
   options: {
     lang: '',
     voiceURI: null,
+    isCloud: null,
     pitch: 1.0,
     rate: 1.0,
     volume: 1
@@ -77,7 +78,8 @@ function speechProviderReducer(state = initialState, action) {
         options: {
           ...state.options,
           voiceURI: action ? action.voiceURI : EMPTY_VOICES,
-          lang: action ? action.lang : DEFAULT_LANG
+          lang: action ? action.lang : DEFAULT_LANG,
+          isCloud: action ? action.isCloud || null : null
         }
       };
     case RECEIVE_TTS_ENGINES:
