@@ -5,12 +5,31 @@ import amber from '@material-ui/core/colors/amber';
 
 import RTLSupport from './RTLSupport';
 
+import { FONTS_FAMILIES, DEFAULT_FONT_FAMILY } from './ThemeProvider.constants';
+
+const setRootFontFamily = fontFamily => {
+  // This is to change fontFamily on all the htmlDocument.
+  // This is to change every textNode that is not contaied in a Typhografy Material UI component.
+  //Considere put all text in a Typhografy.
+  const rootElement = document.querySelector(':root');
+  rootElement.style.fontFamily = fontFamily;
+};
 export class ThemeProvider extends Component {
+  static defaultProps = {
+    fontFamilyName: DEFAULT_FONT_FAMILY
+  };
+
   render() {
-    const { dir, children, darkThemeActive } = this.props;
+    const { dir, children, darkThemeActive, fontFamilyName } = this.props;
+
+    const fontFamily = FONTS_FAMILIES.filter(
+      font => font.fontName === fontFamilyName
+    )[0].fontFamily;
+
+    setRootFontFamily(fontFamily);
 
     const lightTheme = createTheme({
-      typography: {},
+      typography: { fontFamily },
       direction: dir
     });
 
@@ -22,7 +41,7 @@ export class ThemeProvider extends Component {
         },
         secondary: amber
       },
-      typography: {},
+      typography: { fontFamily },
       direction: dir
     });
 
@@ -36,7 +55,8 @@ export class ThemeProvider extends Component {
 
 const mapStateToProps = state => ({
   dir: state.language.dir,
-  darkThemeActive: state.app.displaySettings.darkThemeActive
+  darkThemeActive: state.app.displaySettings.darkThemeActive,
+  fontFamilyName: state.app.displaySettings.fontFamily
 });
 
 const mapDispatchToProps = {};
